+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 All Notable changes to `Slick/di` will be documented in this file
 
-## 2.3.1 - 2017-06-12 [YANKED]
+## 2.4.0 - 2017-12-08
+
+### Removed
+- ``Slick/Common`` dependency
+- ``@inject`` annotation
+- ``ObjectHydratorInterface``s without annotations the hydrators are no longer needed
+- phpunit was unnecessary as all tests are written for phpspec  
+
+## 2.3.1 - 2017-06-12
 
 ### Fixed
 - Resolving dependencies with wrong container in a multi-container scenario was leading to missing